pyproject.toml
[project]
name = "pubmedDataFromQuery"
version = "0.1.0"
description = ""
authors = [
    {name = "Varsha"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "biopython (>=1.85,<2.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
get-papers-list = "pubmedDataFromQuery.main:main"
